clean: refactor misc set difference logic from filter to whereNotIn
What I was trying to accomplish with this piece of code amounted to a set difference method using an accessor function for complex objects. I looked into alternatives and discovered "whereNotIn" and that it, combined with the discovery of Laravel's higher order messages feature, would allow me to cleanly refactor the code.